Make SD and PF incompatible with each other
